+++ stencil/utils.py
@@ -1,0 +1,6 @@
+import sys
+
+
+def abort(message):
+    print >> sys.stderr, 'Error: %s' % message
+    sys.exit(1)